+++ CHANGES.md
@@ -1,6 +1,7 @@
 #2015-07-15
 
 * Fixed a bug in the table in the cost-emissions scatter where it wasn't correctly removing rows
+* Can now display a message of the day
 
 #2015-07-13